requirements.txt
pydantic~=2.9.2
crewai~=0.79.4
python-dotenv~=1.0.1
crewai_poc~=0.1.0
ollama==0.3.3
langchain-ollama==0.2.0
langchain-community~=0.3.4
langchain~=0.3.7